requirements.txt
torch>=1.13.0           # ya da CUDA sürümüne göre: torch==1.13.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
torchvision>=0.14.0
pycocotools>=2.0.6
opencv-python>=4.5.5
Pillow>=9.0.0
matplotlib>=3.5.0
PyYAML>=6.0
tqdm>=4.64.0
fiftyone>=0.20.0
tensorboard
